move portaudio option to main lib; add cmake options for each lib
