+++ libavz/media/include/avz/media.hpp
@@ -4,4 +4,3 @@
 #include <avz/media/FfmpegPopenMedia.hpp>
 #include <avz/media/FfprobeMetadata.hpp>
 #include <avz/media/Media.hpp>
-#include <avz/media/util.hpp>
