+++ libavz/media/src/FfmpegPopenEncoder.cpp
@@ -1,6 +1,6 @@
+#include "util.hpp"
 #include <GL/glew.h>
 #include <avz/media/FfmpegPopenEncoder.hpp>
-#include <avz/media/util.hpp>
 #include <cstring>
 #include <iostream>
 #include <stdexcept>
