+++ libavz/media/src/FfmpegPopenMedia.cpp
@@ -1,5 +1,5 @@
+#include "util.hpp"
 #include <avz/media/FfmpegPopenMedia.hpp>
-#include <avz/media/util.hpp>
 #include <cstring>
 #include <iostream>
