+++ libavz/media/src/FfprobeMetadata.cpp
@@ -1,5 +1,5 @@
+#include "util.hpp"
 #include <avz/media/FfprobeMetadata.hpp>
-#include <avz/media/util.hpp>
 #include <iostream>
 #include <sstream>
