+++ libavz/media/src/util.hpp
